refactor: only code refactor (route, component) (#4213)
* refactor: code refactor


* refactor: code refactor


---------
