+++ apps/web/src/services/preference/routes/admin/routes.ts
@@ -12,6 +12,7 @@ const DomainSettingsPage = () => import('@/services/preference/pages/admin/Admin
 const WorkspacesPage = () => import('@/services/preference/pages/admin/AdminWorkspacesPage.vue');
 const BookmarkContainer = () => import('@/services/preference/components/BookmarkContainer.vue');
 const BookmarkPage = () => import('@/services/preference/pages/admin/AdminBookmarkPage.vue');
+const BookmarkDetailContainer = () => import('@/services/preference/components/BookmarkDetailContainer.vue');
 const BookmarkDetailPage = () => import('@/services/preference/pages/admin/AdminBookmarkDetailPage.vue');
 
 const adminPreferenceRoutes: RouteConfig = {
@@ -53,18 +54,27 @@ const adminPreferenceRoutes: RouteConfig = {
                     component: BookmarkPage,
                 },
                 {
-                    path: ':group',
-                    name: makeAdminRouteName(PREFERENCE_ROUTE.BOOKMARK.GROUP._NAME),
-                    meta: { lsbVisible: true },
+                    path: 'detail',
                     props: true,
-                    component: BookmarkDetailPage,
-                },
-                {
-                    path: ':folder',
-                    name: makeAdminRouteName(PREFERENCE_ROUTE.BOOKMARK.DETAIL._NAME),
-                    meta: { lsbVisible: true },
-                    props: true,
-                    component: BookmarkDetailPage,
+                    component: BookmarkDetailContainer,
+                    children: [
+                        {
+                            path: ':group?',
+                            name: makeAdminRouteName(PREFERENCE_ROUTE.BOOKMARK.DETAIL.GROUP._NAME),
+                            meta: { lsbVisible: true },
+                            props: true,
+                            component: BookmarkDetailPage,
+                            children: [
+                                {
+                                    path: ':folder?',
+                                    name: makeAdminRouteName(PREFERENCE_ROUTE.BOOKMARK.DETAIL.FOLDER._NAME),
+                                    meta: { lsbVisible: true },
+                                    props: true,
+                                    component: BookmarkDetailPage,
+                                },
+                            ],
+                        },
+                    ],
                 },
             ],
         },
