+++ apps/web/src/services/preference/routes/route-constant.ts
@@ -7,8 +7,11 @@ export const PREFERENCE_ROUTE = {
     },
     BOOKMARK: {
         _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}`,
-        GROUP: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}.group` },
-        DETAIL: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}.detail` },
+        DETAIL: {
+            _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}.detail`,
+            GROUP: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}.detail.group` },
+            FOLDER: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.BOOKMARK}.detail.folder` },
+        },
     },
     WORKSPACES: {
         _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.WORKSPACES}`,
